Split hapijs plguins definition
