+++ README.md
@@ -42,6 +42,30 @@ The seed contains a sample [hapi] application (ToDo Lists) and is preconfigured 
   - Mac: http://postgresapp.com/
   - Linux/Windows: http://www.postgresql.org/download/
 
+## Folders structure
+
+```shell
+|-- /config           # Different ENV & CI config files
+|
+|-- /database
+  |-- /migrations     # DB migrations
+  |-- /seeds          # Dummy DB data (seeds)
+|
+|-- /libs
+  |-- /plugins        # hapi.js plugin's
+  |-- /tasks          # Gulp tasks
+|
+|-- /public           # Public accessible content
+|
+|-- /src              # API: here is where you want to add your code
+  |-- /controllers    # Route handlers
+  |-- /models         # Business logic
+  |-- /routes         # API RESTful end-points
+  |-- /services       # 3rd. party services (Slack, Facebook, etc.)
+|
+|-- /tests
+  |-- /api            # API & Integration tests
+```
 
 ## Getting started
