+++ libs/plugins/README.md
@@ -1,0 +1,1 @@
+## hapi.js plugins
