+++ package.json
@@ -30,7 +30,6 @@
     "good-file": "^6.0.1",
     "good-squeeze": "^4.0.0",
     "hapi": "^13.4.1",
-    "hapi-auth-jwt2": "^7.0.1",
     "inert": "^4.0.0",
     "joi": "^8.0.0",
     "knex": "^0.11.0",
