+++ src/services/README.md
@@ -1,1 +1,1 @@
-## Save code here that interacts with 3rd. party tools. Such as Slack, Dropbox, Facebook, etc.
+## Put code here that interacts with 3rd. party tools. Such as Slack, Dropbox, Facebook, etc.
