+++ tests/helpers.js
@@ -1,5 +1,5 @@
 import _  from 'lodash'
-import { default as log } from '../src/logger'
+import { default as log } from '../libs/logger'
 
 
 //
